# modules/dxp/apps/portal-workflow/portal-workflow-metrics-rest-impl/rest-openapi.yaml
components:
    schemas:
...
        Task:
            description:
                https://www.schema.org/Task
            properties:
                assigneeId:
                    format: int64
                    type: integer
                className:
                    type: string
                classPK:
                    format: int64
                    type: integer
                completed:
                    type: boolean
                completionUserId:
                    format: int64
                    type: integer
                dateCompletion:
                    format: date-time
                    type: string
                dateCreated:
                    format: date-time
                    type: string
                dateModified:
                    format: date-time
                    type: string
                duration:
                    format: int64
                    type: integer
                id:
                    format: int64
                    type: integer
                instanceId:
                    format: int64
                    type: integer
                label:
                    readOnly: true
                    type: string
                name:
                    type: string
                nodeId:
                    format: int64
                    type: integer
                processId:
                    format: int64
                    type: integer
                processVersion:
                    type: string
            type: object
...
info:
    description:
        ""
    license:
        name: "Apache 2.0"
        url: "http://www.apache.org/licenses/LICENSE-2.0.html"
    title: ""
    version: v1.0
openapi: 3.0.1
paths:
...
    "/processes/{processId}/tasks":
        get:
            operationId: getProcessTasksPage
            parameters:
                - in: path
                  name: processId
                  required: true
                  schema:
                      format: int64
                      type: integer
                - in: query
                  name: page
                  schema:
                      type: integer
                - in: query
                  name: pageSize
                  schema:
                      type: integer
                - in: query
                  name: sort
                  schema:
                      type: string
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                items:
                                    $ref: "#/components/schemas/Task"
                                type: array
                        application/xml:
                            schema:
                                items:
                                    $ref: "#/components/schemas/Task"
                                type: array
                    description:
                        ""
            tags: ["Task"]
        post:
            operationId: postProcessTask
            parameters:
                - in: path
                  name: processId
                  required: true
                  schema:
                      format: int64
                      type: integer
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/Task"
                    application/xml:
                        schema:
                            $ref: "#/components/schemas/Task"
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Task"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/Task"
                    description:
                        ""
            tags: ["Task"]
...
    "/processes/{processId}/tasks/{taskId}":
        delete:
            operationId: deleteProcessTask
            parameters:
                - in: path
                  name: processId
                  required: true
                  schema:
                      format: int64
                      type: integer
                - in: path
                  name: taskId
                  required: true
                  schema:
                      format: int64
                      type: integer
            responses:
                204:
                    content:
                        application/json: {}
                        application/xml: {}
                    description:
                        ""
            tags: ["Task"]
        get:
            operationId: getProcessTask
            parameters:
                - in: path
                  name: processId
                  required: true
                  schema:
                      format: int64
                      type: integer
                - in: path
                  name: taskId
                  required: true
                  schema:
                      format: int64
                      type: integer
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Task"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/Task"
                    description:
                        ""
            tags: ["Task"]
        patch:
            operationId: patchProcessTask
            parameters:
                - in: path
                  name: processId
                  required: true
                  schema:
                      format: int64
                      type: integer
                - in: path
                  name: taskId
                  required: true
                  schema:
                      format: int64
                      type: integer
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/Task"
                    application/xml:
                        schema:
                            $ref: "#/components/schemas/Task"
            responses:
                204:
                    content:
                        application/json: {}
                        application/xml: {}
                    description:
                        ""
            tags: ["Task"]
...